--- src/helpers/createElement.js
@@ -1,6 +1,0 @@
-const templateElem = document.createElement(`template`);
-
-export const createElement = (str) => {
-  templateElem.innerHTML = str;
-  return templateElem.content.firstElementChild;
-};
