+++ src/helpers/get-class.js
