+++ src/helpers/get-films-by-filter.js
