+++ src/helpers/get-list-as-str.js
