+++ src/helpers/index.js
@@ -1,16 +1,10 @@
-export {createElement} from './createElement';
-export {renderElement} from './renderElement';
-export {removeElement} from './removeElement';
-export {replaceElement} from './replaceElement';
-export {getClass} from './getClass';
-export {getDate, getFullDate, getRelativeDate, getDuration} from './getDateTime';
-export {getFilmControlsData} from './getFilmControlsData';
-export {getPlurals} from './getPlurals';
-export {getListAsStr} from './getListAsStr';
-export {getFilmsByFilter} from './getFilmsByFilter';
+export { createElement, renderElement, removeElement, replaceElement} from './dom';
+export {getClass} from './get-class';
+export { getDate, getFullDate, getRelativeDate, getDuration } from './date-time';
+export {getPlurals} from './get-plurals';
+export {getListAsStr} from './get-list-as-str';
+export {getFilmsByFilter} from './get-films-by-filter';
 export {sortByRating, sortByDate, sortByComments, getFilmsSortedByProp} from './sort';
-export {getHandlerWithProp} from './getHandlerWithProp';
-export {getHandlerWithValue} from './getHandlerWithValue';
-export {getHandlerToggleClass} from './getHandlerToggleClass';
+export { getHandlerWithProp, getHandlerWithValue, getHandlerToggleClass} from './handlers';
 export {getWatched, getTotalDuration, getWatchedByGenre, getUserStatus, getWatchedByPeriod} from './stats';
 export {shake} from './shake';
