--- src/helpers/replaceElement.js
@@ -1,7 +1,0 @@
-export const replaceElement = (oldComponent, newComponent) => {
-  if (!oldComponent && !newComponent) {
-    return;
-  }
-
-  oldComponent.getElement().replaceWith(newComponent.getElement());
-};
